requirements.txt
poe-api-wrapper~=1.3.3
pydantic-mongo @ git+https://github.com/laruss/pydantic-mongo
python-decouple~=3.8
flask-cors~=4.0.0
python-dateutil~=2.8.2
pydantic~=2.4.2
flask~=3.0.0
flask_swagger_ui~=4.11.1
